Improved click events and ways to show/hide button
Buttons can now be shown and hidden not only based on the URL of the page shown but also based on content within that page. One can eg. look for a specific kind of link relations in the page's head tag and depending on the value of those links decide to hide or show the button.

The click events has been improved so that they now are called with the context set to the document of the HTML page. This means that if your script has forgotten why it showed the button then it can recheck the reason. This is important when privacy mode is activated as caching results then can be a bit problematic.